refactor: user 도메인 member로 변경
