+++ src/main/java/com/amcamp/cineAI/domain/auth/application/AuthService.java
@@ -3,9 +3,9 @@ package com.amcamp.cineAI.domain.auth.application;
 import com.amcamp.cineAI.domain.auth.dto.request.AuthCodeRequest;
 import com.amcamp.cineAI.domain.auth.dto.response.MemberInfoResponse;
 import com.amcamp.cineAI.domain.auth.dto.response.SocialLoginResponse;
-import com.amcamp.cineAI.domain.user.dao.MemberRepository;
-import com.amcamp.cineAI.domain.user.domain.Member;
-import com.amcamp.cineAI.domain.user.domain.MemberStatus;
+import com.amcamp.cineAI.domain.member.dao.MemberRepository;
+import com.amcamp.cineAI.domain.member.domain.Member;
+import com.amcamp.cineAI.domain.member.domain.MemberStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
