+++ src/main/java/com/amcamp/cineAI/domain/member/domain/MemberStatus.java
@@ -1,4 +1,4 @@
-package com.amcamp.cineAI.domain.user.domain;
+package com.amcamp.cineAI.domain.member.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
